# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/chirps": {
            "get": {
                "description": "Getting all Chirps if query string is empty or all Chirps by the author if authorID is provided in query string, you also can provide sort order in query string",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chirps"
                ],
                "summary": "Getting all Chirps or all Chirps by the author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorID",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order, can either be 'asc' or 'desc', asc by default",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ChirpValidRespBody"
                            }
                        }
                    }
                }
            }
        },
        "/api/chirps/{chirpID}": {
            "get": {
                "description": "finding chirp by chirpID provided in path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chirps"
                ],
                "summary": "finding chirp by chirpID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "chirpID",
                        "name": "chirpID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ChirpValidRespBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ChirpValidRespBody": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}